requirements.txt
asgiref==3.11.1
charset-normalizer==3.4.5
Django==6.0.3
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
factory_boy==3.3.3
Faker==40.11.0
gunicorn==25.1.0
iniconfig==2.3.0
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.12.1
pytest==9.0.2
pytest-django==4.12.0
phonenumbers==8.13.55
python-decouple==3.8
reportlab==4.4.10
sqlparse==0.5.5
